Refactor single `fields` module to mulit-modules
- will be a lot simpler to extend and change little parts if not all in one module
