+++ src/search/search-index-new/storage/fields/field.ts
@@ -1,0 +1,9 @@
+export class Field {
+    prepareForStorage(input) {
+        return input
+    }
+
+    prepareFromStorage(stored) {
+        return stored
+    }
+}
