+++ src/search/search-index-new/storage/fields/index.ts
@@ -1,0 +1,7 @@
+import { UrlField } from './url-field'
+
+export { Field } from './field'
+
+export default {
+    url: UrlField,
+}
